Change definition of isolated pawns
We now consider to be isolated a pawn which has no friendly pawn in the
three squares on its left or right sides : so an isolated pawn is not
supported, not part of a phalanx and not supporting any other pawn.
Divide the Isolated[] penalty by two as there are many more isolated
pawns by this definition.