Bonus for pins on the opponent queen (exclude batteries)
Bonus for pins on the opponent queen (take 2).

This version excludes from the pinned pieces rook blockers which are
orthogonally pinned and bishop blockers which are diagonally pinned, as
these pieces are in fact part of a battery with the queen and thus may
be both a weakness and a strength.