queenless endgame love = 15
Increase scale factor for the strong size, decrease it for the weak
side in evaluate_scale_factor().

The obvious effect will be that Stockfish will love entering queenless
endgames if they are favorable, and fear them less is they are
unfavorable.

The side effect should theoretically be that SF keeps more minor pieces
and rooks, relatively to current master.

Bench = 7645444