Fix division by PHASE_MIDGAME
Fix division by PHASE_MIDGAME in evaluate().

Before the patch, since PHASE_MIDGAME is 128, we had the following
mapping of values by the integer division by PHASE_MIDGAME in line 832 :

[-383 .. -256]  => -2
[-255 .. -128]  =>  -1
[-127..127]      =>  0
[128 .. 255]      => 1
[256 .. 383]      => 2

We see that the central interval, which was mapped to 0, was twice as
large as the other ones. The patch corrects this, such that the
intervals mapped to consecutive integers have all the same width :

[-319 .. -192]  => -2
[-191 .. -64]  =>  -1
[-63..63]         =>  0
[64 .. 191]       => 1
[192 .. 319]      => 2