Use game phases outside 0..128
The game phase is 0 (PHASE_ENDGAME) when the non-pawn material has value EndgameLimit (about rook and knight for each player), and 128 (PHASE_MIDGAME) when the non-pawn material has value MidgameLimit (about queen, 2 rooks, 2 bishops and 1 knight for each player).

With this patch the phase can be outside the range 0..128 for positions with even less or more material than these limits, thus extrapolating a little bit more the tapered eval.