Change for sudden death
When playing in sudden death mode (increment = 0), we allocate time as if we had to play at least 10 more future moves after *each* move: in other words, we shall never use more than 1/10th of the remaining time for each single move. We also amplify the moveOverhead value to anticipate the sum of overheads for the next ten moves.