hook = S(5,5)
A hook is a pawn of our color, usually advanced, which is challenged and which will force a defect in our pawn structure no matter how we defend. Here we consider as hooks the levers which give a passed pawn to the opponent or create isolated pawn(s) in our camp if we execute the capture.

Bench: 4630411