Back to the drawing board
Back to the drawing board. The patches have trouble passing LTC so I use a more complicated matrix, but with the advantage that the pruning formula will be 100% equivalent to master when the matrix is not used. This will (hopefully) help the patches to be on par with master when I exclude higher depths.

Bench: 4409049