No scale_factor() call when queens are on
Fast exit of evaluate::scale_factor(), returning SCALE_FACTOR_NORMAL, when there is at least one queen on the board. Testing with [0..5] bounds. This may be slightly faster overall, but also a bit controversial since this simplify away the fortress detection of Q vs R+Pawns  with a rook on the third rank defended by a pawn in endgame function KQKRPs.

Bench: 4017007