Evaluations of Vancura endgame positions with Black's rook on a1
This patch allows Stockfish to find the objective evaluations for a rook and pawn vs rook endgame where
White’s pawn is on a6, their rook is in front of it (a7/a8), and Black’s rook is behind it. A few more
specific conditions have to be met as well (see the outer if statement), so for now this feature cannot
return an objective evaluation for all Vancura positions with the rook on the a-file.

The feature is a bit over 40 lines, but it is all contained at the end of one function in endgames.cpp that
isn't modified much. As for efficiency, it shouldn't cause any issues since before the feature's
code can even run, the conditions in the outer if statement have to be met. Each of the conditions is also
roughly ordered by their likelihood of being false, in order to benefit from short circuit evaluation.