Introduce USE_POSIX_ALIGNED_ALLOC compilation flag
With this patch, it is now possible to pass USE_POSIX_ALIGNED_ALLOC to
the compiler flags to force the use of the posix_memalign() function
as aligned memory allocator.

The behavior is unchanged compared to master if USE_POSIX_ALIGNED_ALLOC
if not defined, but if USE_POSIX_ALIGNED_ALLOC is defined we now fall
back to posix_memalign() if available. This allows the compilation of
Stockfish on Macintosh prior to 10.15 (using the clang compiler).

====

Example of usage :

export CXXFLAGS=-DUSE_POSIX_ALIGNED_ALLOC
make -j profile-build COMP=clang COMPCXX=clang++-mp-11

====

closes https://github.com/official-stockfish/Stockfish/pull/5426

No functional change